Updated so that oceanmask is used instead of minimum depth

+++ branches/o2_hindcast/src/masks.c
@@ -9,7 +9,7 @@ extern double umask[NXMEM][NYMEM];   /* _mask are 1 over ocean and 0  */
 extern double vmask[NXMEM][NYMEM];   /* over land on the u & v grids. */
 extern double oceanmask[NXMEM][NYMEM];
 
-extern double D[NXMEM][NYMEM];
+extern int D[NXMEM][NYMEM];
 
 void initializemasks() {
 
@@ -37,9 +37,9 @@ void initializemasks() {
   for(i=0;i<NXMEM;i++)
 	for(j=0;j<NYMEM;j++)
 	if (D[NXMEM][NYMEM]<=Dmin)
-		oceanmask[i][j] = 0.0;
+		oceanmask[i][j] = 0;
 	else
-		oceanmask[i][j] = 1.0;
+		oceanmask[i][j] = 1;
 
   /*
   for (j=27;j<=33;j++) {
